Fix issue for high resolutions

+++ Assets/Scripts/_Utils/LibMesh/MonoBehaviours/Streams/AdvancedMultiStreamProceduralMesh.cs
@@ -23,10 +23,10 @@ public class AdvancedMultiStreamProceduralMesh : MonoBehaviour {
 			VertexAttribute.Normal, dimension: 3, stream: 1
 		);
 		vertexAttributes[2] = new VertexAttributeDescriptor(
-			VertexAttribute.Tangent, VertexAttributeFormat.Float16, 4, 2
+			VertexAttribute.Tangent, VertexAttributeFormat.Float32, 4, 2
 		);
 		vertexAttributes[3] = new VertexAttributeDescriptor(
-			VertexAttribute.TexCoord0, VertexAttributeFormat.Float16, 2, 3
+			VertexAttribute.TexCoord0, VertexAttributeFormat.Float32, 2, 3
 		);
 		meshData.SetVertexBufferParams(vertexCount, vertexAttributes);
 		vertexAttributes.Dispose();
@@ -40,20 +40,20 @@ public class AdvancedMultiStreamProceduralMesh : MonoBehaviour {
 		NativeArray<float3> normals = meshData.GetVertexData<float3>(1);
 		normals[0] = normals[1] = normals[2] = normals[3] = back();
 
-		half h0 = half(0f), h1 = half(1f);
+		float h0 = 0f, h1 = 1f;
 
-		NativeArray<half4> tangents = meshData.GetVertexData<half4>(2);
+		NativeArray<float4> tangents = meshData.GetVertexData<float4>(2);
 		tangents[0] = tangents[1] = tangents[2] = tangents[3] =
-			half4(h1, h0, h0, half(-1f));
+			float4(h1, h0, h0, -1f);
 		
-		NativeArray<half2> texCoords = meshData.GetVertexData<half2>(3);
+		NativeArray<float2> texCoords = meshData.GetVertexData<float2>(3);
 		texCoords[0] = h0;
-		texCoords[1] = half2(h1, h0);
-		texCoords[2] = half2(h0, h1);
+		texCoords[1] = float2(h1, h0);
+		texCoords[2] = float2(h0, h1);
 		texCoords[3] = h1;
 
-		meshData.SetIndexBufferParams(triangleIndexCount, IndexFormat.UInt16);
-		NativeArray<ushort> triangleIndices = meshData.GetIndexData<ushort>();
+		meshData.SetIndexBufferParams(triangleIndexCount, IndexFormat.UInt32);
+		NativeArray<uint> triangleIndices = meshData.GetIndexData<uint>();
 		triangleIndices[0] = 0;
 		triangleIndices[1] = 2;
 		triangleIndices[2] = 1;
